Added netdrive, logmode, domainuser, startmode and workingdir
- ENV expansion performed throughout - added 'workingdir' to specify the
runtime working directory of the service. - added 'domainuser' to
specify a username when installing the service.  Prompts for password
during installation.  - added 'startmode' to be used when installing the
service. - added 'netdrive' to instruct the service wrapper to mount the
shared drive at Service startup and unmount the shared drives on Service
shutdown. - enhanced 'env' configuration to support append and prepend
for path building - added 'logdir' to specify where log files are to be
generated.  Directory created if not present. - made handling of
'arguments', 'startarguments' and 'stoparguments' uniform. - added
command line feedback for error conditions.