Update Main.cs
When stopping the children processes of a parent process, do not stop a child if it is executing the service wrapper itself (for instance, because the parent spawned a process to execute "[wrapper.exe] restart"). This is a resolution for [JENKINS-22685].

Also, updated logging of SIGINT to properly reflect the PID being stopped. Previously, it was always printing the main parent's PID even when SIGINTing a child PID.